implement zarr backend

+++ pyanndata/src/data/slice.rs
@@ -1,3 +1,5 @@
+use std::ffi::c_long;
+
 use crate::data::instance::*;
 
 use pyo3::prelude::*;
@@ -18,7 +20,7 @@ pub fn to_select_info(ob: &Bound<'_, PyAny>, shape: &Shape) -> PyResult<SelectIn
 
 pub fn to_select_elem(ob: &Bound<'_, PyAny>, length: usize) -> PyResult<SelectInfoElem> {
     let select = if let Ok(slice) = ob.downcast::<pyo3::types::PySlice>() {
-        let s = slice.indices(length as isize)?;
+        let s = slice.indices(length as c_long)?;
         ndarray::Slice { 
             start: s.start,
             end: Some(s.stop),
